pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "anthropic>=0.52.0",
    "email-validator>=2.2.0",
    "flask>=3.1.1",
    "flask-sqlalchemy>=3.1.1",
    "gunicorn>=23.0.0",
    "openai>=1.82.0",
    "opencv-python>=4.11.0.86",
    "openpyxl>=3.1.5",
    "pillow>=11.2.1",
    "psycopg2-binary>=2.9.10",
    "pypdf2>=3.0.1",
    "pytesseract>=0.3.13",
    "python-docx>=1.1.2",
    "python-telegram-bot>=22.1",
    "pyyaml>=6.0.2",
    "requests>=2.32.3",
    "sqlalchemy>=2.0.41",
    "werkzeug>=3.1.3",
]
